requirements.txt
tronpy
requests 
python-dotenv